serving-service/requirements.txt
# --- ADK Core & Web Server ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9  # Required for file uploads

# --- Async HTTP (Agent-to-Agent Comm) ---
aiohttp>=3.9.3

# --- Google Cloud AI ---
google-cloud-aiplatform>=1.43.0  # Vertex AI SDK
vertexai>=1.43.0                 # Generative Models

# --- Google Cloud Data & Storage ---
google-cloud-discoveryengine>=0.11.0  # Vertex AI Search
google-cloud-firestore>=2.15.0        # Content Store
google-cloud-storage>=2.14.0          # Image Blobs

# --- Utilities ---
python-dotenv>=1.0.1  # For local development